Extracting integration tests in submodule
