+++ core-integration-tests/build.gradle
@@ -22,31 +22,11 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.test;
+apply plugin: 'java'
 
-import com.jaspersoft.android.sdk.network.RestApiLog;
-
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
-/**
- * @author Tom Koptel
- * @since 2.0
- */
-public final class TestLogger implements RestApiLog {
-
-    private final Logger logger;
-
-    private TestLogger(String logTarget) {
-        logger = Logger.getLogger(logTarget);
-    }
-
-    public static TestLogger get(Object target) {
-        return new TestLogger(target.getClass().getSimpleName());
-    }
-
-    @Override
-    public void log(String message) {
-        logger.log(Level.INFO, message);
-    }
+dependencies {
+    compile project(':js-android-sdk-core')
+    testCompile 'org.hamcrest:hamcrest-integration:1.3'
+    testCompile 'junit:junit:4.12'
+    testCompile 'org.bouncycastle:bcprov-jdk16:1.46'
 }
