+++ core-integration-tests/src/test/java/com/jaspersoft/android/sdk/network/AuthenticationRestApiTest.java
@@ -22,12 +22,12 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.test.integration.api;
+package com.jaspersoft.android.sdk.network;
 
 import com.jaspersoft.android.sdk.network.AuthenticationRestApi;
 import com.jaspersoft.android.sdk.network.JSEncryptionAlgorithm;
 import com.jaspersoft.android.sdk.network.entity.server.EncryptionKey;
-import com.jaspersoft.android.sdk.test.TestLogger;
+import com.jaspersoft.android.sdk.util.TestLogger;
 
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.junit.Ignore;
