+++ core-integration-tests/src/test/java/com/jaspersoft/android/sdk/network/ReportOptionRestApiTest.java
@@ -22,13 +22,13 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.test.integration.api;
+package com.jaspersoft.android.sdk.network;
 
 import com.jaspersoft.android.sdk.network.ReportOptionRestApi;
 import com.jaspersoft.android.sdk.network.entity.report.option.ReportOption;
-import com.jaspersoft.android.sdk.test.TestLogger;
-import com.jaspersoft.android.sdk.test.integration.api.utils.DummyTokenProvider;
-import com.jaspersoft.android.sdk.test.integration.api.utils.JrsMetadata;
+import com.jaspersoft.android.sdk.util.DummyTokenProvider;
+import com.jaspersoft.android.sdk.util.JrsMetadata;
+import com.jaspersoft.android.sdk.util.TestLogger;
 
 import org.junit.Before;
 import org.junit.Test;
