+++ core-integration-tests/src/test/java/com/jaspersoft/android/sdk/util/DummyTokenProvider.java
@@ -22,8 +22,42 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-include ':js-android-sdk-client'
-include ':js-android-sdk-core'
-include ':core-integration-tests'
-project(':js-android-sdk-client').projectDir = "$rootDir/client" as File
-project(':js-android-sdk-core').projectDir = "$rootDir/core" as File
+package com.jaspersoft.android.sdk.util;
+
+import com.jaspersoft.android.sdk.network.AuthenticationRestApi;
+
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * @author Tom Koptel
+ * @since 2.0
+ */
+public final class DummyTokenProvider {
+
+    private final JrsMetadata mJrsMetadata;
+    private String mToken;
+
+    public DummyTokenProvider(JrsMetadata jrsMetadata) {
+        mJrsMetadata = jrsMetadata;
+    }
+
+    public static DummyTokenProvider create(JrsMetadata metadata) {
+        return new DummyTokenProvider(metadata);
+    }
+
+    @NotNull
+    public String provideToken() {
+        if (mToken == null) {
+            AuthenticationRestApi restApi = new AuthenticationRestApi.Builder()
+                    .baseUrl(mJrsMetadata.getServerUrl())
+                    .build();
+            mToken = restApi
+                    .authenticate(mJrsMetadata.getUsername(), mJrsMetadata.getPassword(), mJrsMetadata.getOrganization(), null);
+        }
+        return mToken;
+    }
+
+    public String token() {
+        return provideToken();
+    }
+}
